requirements.txt
djangorestframework==3.14.0
psycopg2-binary==2.9.5
Django==3.2.16